Split off bokeh.resources.sri
