+++ setup.cfg
@@ -98,8 +98,15 @@ warn_unused_configs = True
 [mypy-bokeh.*]
 ignore_errors = True
 
-[mypy-bokeh.resources]
+[mypy-tests.*]
+ignore_errors = True
+
+[mypy-bokeh.resources.*]
 ignore_errors = False
+
+# Currently ignored due to lack of tests/unit/**/__init__.py
+[mypy-tests.unit.bokeh.resources.*]
+ignore_errors = True
 
 [mypy-bokeh.palettes]
 ignore_errors = False
